Add files via upload
This follows Carly's demo on adding a search bar. I was able to customize the search bar to sort of match the one we had using bootstrap, but not quite. It also uses another API (Geoapify) as I'm not able to add a custom search bar using the previous API in index.html.